Create ArrayClone
String[] java.lang.Object.clone()


Creates and returns a copy of this object. The precise meaningof "copy" may depend on the class of the object. The generalintent is that, for any object x, the expression: 

 x.clone() != x
will be true, and that the expression: 
 x.clone().getClass() == x.getClass()
will be true, but these are not absolute requirements.While it is typically the case that: